Added DBFacade.cs
Simple implementation of SQLite Database. As of now, remember to run dump.sql and schema.sql before first run.

#20
