+++ src/Chirp.Razor/DBFacade.cs
@@ -1,83 +1,67 @@
+namespace Chirp.Razor;
 using Microsoft.Data.Sqlite;
-using System;
-using System.Collections.Generic;
-using System.IO;
-
-namespace DefaultNamespace;
-
 public class DBFacade
 {
-    private readonly string sqlDBFilePath;
+    private readonly string _connectionString;
+
     public DBFacade()
     {
         var dbPath = Environment.GetEnvironmentVariable("CHIRPDBPATH");
-
-        if (string.IsNullOrEmpty(dbPath)) {
-            dbPath = Path.Combine(Path.GetTempPath(), "chirp.db");
+        if (string.IsNullOrWhiteSpace(dbPath))
+        {
+            var temp = Path.GetTempPath();
+            dbPath = Path.Combine(temp, "chirp.db");
         }
-        
-        sqlDBFilePath = $"Data Source={dbPath}";
+        _connectionString = $"Data Source={dbPath}";
     }
 
     public List<CheepViewModel> GetCheeps()
     {
-        var cheeps = new List<CheepViewModel>();
+        using var conn = new SqliteConnection(_connectionString);
+        conn.Open();
 
-        using var connection  = new SqliteConnection(sqlDBFilePath);
-        connection.Open();
-        
-        var sqlQuery = @"
-                SELECT u.username, m.text, m.pub_date
-                FROM message m
-                JOIN user u ON m.author_id = u.user_id
-                ORDER BY m.pub_date DESC;";
-
-        
-        var command = connection.CreateCommand();
-        command.CommandText = sqlQuery;
-
-        using var reader = command.ExecuteReader();
-        while (reader.Read())
-        {
-            cheeps.Add(new CheepViewModel(
-                reader.GetString(0), // Author
-                reader.GetString(1), // Message 
-                reader.GetString(2)  // Timestamp
-            ));
-        }
-
-        return cheeps;
+        using var cmd = conn.CreateCommand();
+        cmd.CommandText = @"
+            SELECT u.username, m.text, m.pub_date
+            FROM message m
+            JOIN user u ON m.author_id = u.user_id
+            ORDER BY m.pub_date DESC";
+        return ReadCheeps(cmd);
     }
 
     public List<CheepViewModel> GetCheepsFromAuthor(string author)
     {
-        var cheeps = new List<CheepViewModel>();
-        
-        using var connection  = new SqliteConnection(sqlDBFilePath);
-        connection.Open();
-        
-        var sqlQuery = @"
-                SELECT u.username, m.text, m.pub_date
-                FROM message m
-                JOIN user u ON m.author_id = u.user_id
-                WHERE u.username = @author
-                ORDER BY m.pub_date DESC;";
-        
-        var command = connection.CreateCommand();
-        command.CommandText = sqlQuery;
-        command.Parameters.AddWithValue("@author", author);
+        using var conn = new SqliteConnection(_connectionString);
+        conn.Open();
 
-        using var reader = command.ExecuteReader();
-        
-        while (reader.Read())
-        {
-            cheeps.Add(new CheepViewModel(
-                reader.GetString(0), // Author
-                reader.GetString(1), // Message 
-                reader.GetString(2)  // Timestamp
-            ));
-        }
-
-        return cheeps;
+        using var cmd = conn.CreateCommand();
+        cmd.CommandText = @"
+            SELECT u.username, m.text, m.pub_date
+            FROM message m
+            JOIN user u ON m.author_id = u.user_id
+            WHERE u.username = @author
+            ORDER BY m.pub_date DESC";
+        cmd.Parameters.AddWithValue("@author", author);
+        return ReadCheeps(cmd);
     }
+
+    private static List<CheepViewModel> ReadCheeps(SqliteCommand cmd)
+    {
+        var list = new List<CheepViewModel>();
+        using var r = cmd.ExecuteReader();
+        while (r.Read())
+        {
+            var author = r.GetString(0);
+            var message = r.GetString(1);
+            var ts = r.GetInt64(2);
+            list.Add(new CheepViewModel(author, message, UnixToLocal(ts)));
+        }
+        return list;
+    }
+
+    private static string UnixToLocal(long unixSeconds) =>
+        DateTimeOffset.FromUnixTimeSeconds(unixSeconds)
+            .ToLocalTime()
+            .ToString("MM/dd/yy H:mm:ss");
+    
 }
