+++ src/Chirp.Razor/Program.cs
@@ -7,6 +7,7 @@ Batteries.Init();
 
 // Add services to the container.
 builder.Services.AddRazorPages();
+builder.Services.AddSingleton<DBFacade>();
 builder.Services.AddSingleton<ICheepService, CheepService>();
 
 var app = builder.Build();
